+++ tests/test_serialization.py
@@ -9,6 +9,7 @@ import pytest
 from djmoney.money import Money
 from djmoney.serializers import Deserializer, Serializer
 
+from ._compat import patch
 from .testapp.models import ModelWithDefaultAsInt
 
 
@@ -78,3 +79,18 @@ def test_old_fields_skip(capsys, fixture_file):
 def test_deserialization_error():
     with pytest.raises(DeserializationError):
         list(Deserializer("invalid JSON"))
+
+
+def test_patched_get_model(fixture_file):
+    """Sometimes it is useful to patch `django.code.serializers.python._get_model`.
+
+    Our code should use the patched version."""
+    data = '[{"model": "testapp.unknown_model", "pk": 1, "fields": {"money_currency": "USD", "money": "1.00"}}]'
+    fixture_file.write(data)
+
+    def _get_model(identifier):
+        return ModelWithDefaultAsInt
+
+    with patch('django.core.serializers.python._get_model', _get_model):
+        loaddata(fixture_file)
+    assert ModelWithDefaultAsInt.objects.get().money == Money(1, "USD")